requirements.txt
blinker==1.9.0
certifi==2024.7.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
gunicorn==22.0.0
itsdangerous==2.2.0
idna==3.7
Jinja2==3.1.6
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2